+++ tests/TestCase.php
@@ -13,19 +13,28 @@ abstract class TestCase extends BaseTestCase
      * @param array<string, int|string> $headers
      */
     protected function mockResponse(
-        string $mockfile = 'test',
+        string $mockfile = "test",
         int $statusCode = 200,
         array $headers = [],
     ): MockResponse {
-        $path = sprintf('./tests/Feature/Data/%s.json', $mockfile);
-        $content = file_get_contents($path);
-        if ($content === false) {
-            throw new \RuntimeException(sprintf('Failed to read mock file: %s', $path));
-        }
+        $content = $this->mockData($mockfile);
 
         return new MockResponse($content, [
-            'http_code' => $statusCode,
-            'response_headers' => $headers,
+            "http_code" => $statusCode,
+            "response_headers" => $headers,
         ]);
+    }
+
+    protected function mockData(string $mockfile = "test"): string
+    {
+        $path = sprintf("./tests/Feature/Data/%s.json", $mockfile);
+        $content = file_get_contents($path);
+        if ($content === false) {
+            throw new \RuntimeException(
+                sprintf("Failed to read mock file: %s", $path),
+            );
+        }
+
+        return $content;
     }
 }